fixed error with plugin not working with ckan 2.9, serparated function in controller.py to be in a util file and have blueprint + controller refer to utils, created a directory for 2.7 controller, and delted controllers.py
